# back/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gonglijeongli API",
    "description": "API Documentation for Gonglijeongli Backend",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative URL for current domain"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Dev Server"
    }
  ],
  "paths": {
...
    "/api/events/": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "행사 목록 조회 (페이지네이션)",
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "페이지당 항목 수",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "새 행사 생성",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "행사 데이터",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Sample Event"
                },
                "date": {
                  "type": "string",
                  "example": "2023-10-31"
                },
                "location": {
                  "type": "string",
                  "example": "Seoul"
                },
                "thumb": {
                  "type": "string",
                  "example": "https://example.com/image.jpg"
                },
                "link": {
                  "type": "string",
                  "example": "https://example.com"
                },
                "status": {
                  "type": "string",
                  "example": "upcoming"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "date": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  },
                  "thumb": {
                    "example": "any"
                  },
                  "link": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{id}": {
      "put": {
        "tags": [
          "Events"
        ],
        "summary": "행사 수정",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "행사 ID"
          },
          {
            "name": "body",
            "in": "body",
            "description": "수정할 행사 데이터",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Updated Event"
                },
                "date": {
                  "type": "string",
                  "example": "2023-11-01"
                },
                "location": {
                  "type": "string",
                  "example": "Busan"
                },
                "thumb": {
                  "type": "string",
                  "example": "https://example.com/updated.jpg"
                },
                "link": {
                  "type": "string",
                  "example": "https://example.com/updated"
                },
                "status": {
                  "type": "string",
                  "example": "ongoing"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "date": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  },
                  "thumb": {
                    "example": "any"
                  },
                  "link": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Events"
        ],
        "summary": "행사 삭제",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "행사 ID"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    }
  }
}